+++ src/abstractruntime/storage/__init__.py
@@ -3,6 +3,12 @@
 from .base import DeletableLedgerStore, DeletableRunStore, LedgerStore, QueryableRunStore, RunStore
 from .in_memory import InMemoryRunStore, InMemoryLedgerStore
 from .json_files import JsonFileRunStore, JsonlLedgerStore
+from .lease import (
+    DirectoryLease,
+    DirectoryLeaseHeld,
+    acquire_directory_lease,
+    read_directory_lease,
+)
 from .ledger_chain import HashChainedLedgerStore, verify_ledger_chain
 from .observable import ObservableLedgerStore, ObservableLedgerStoreProtocol
 from .offloading import OffloadingLedgerStore, OffloadingRunStore, offload_large_values
@@ -18,6 +24,10 @@ __all__ = [
     "InMemoryLedgerStore",
     "JsonFileRunStore",
     "JsonlLedgerStore",
+    "DirectoryLease",
+    "DirectoryLeaseHeld",
+    "acquire_directory_lease",
+    "read_directory_lease",
     "HashChainedLedgerStore",
     "verify_ledger_chain",
     "ObservableLedgerStore",